Minor refactoring and html changes
Fixed html typos, added homepage links, set JS to remove Home link from iframes, and  added media queries for full page maps

+++ Flask_Application/application/models_Strava.py
@@ -28,7 +28,7 @@ class sub_update(Base):
     id = Column(Integer, primary_key=True)
     aspect = Column(String)
     event_time = Column(DateTime)
-    object_id = Column(Integer)
+    object_id = Column(BigInteger)
     object_type = Column(String)
     owner_id = Column(Integer)
     subscription_id = Column(Integer)
